More work on the Yahoo parser. Minor fixes to AJAX code.
darcs-hash:20060913182402-f1522-d44d88070b24605657304d7c3fdc1a78ca2fd7c5.gz

+++ src/plugins/NetworkForensics/ProtocolHandlers/MSN.py
@@ -1391,7 +1391,7 @@ class MSNScanner(StreamScannerFactory):
             """ CREATE TABLE if not exists `msn_session` (
             `inode` VARCHAR(50) NOT NULL,
             `packet_id` INT NOT NULL,
-            `session_id` INT,
+            `session_id` BIGINT,
             `sender` VARCHAR(250),
             `recipient` VARCHAR( 250 ),
             `type` VARCHAR(50),
